user API 수정

+++ backend/routes/api/user/index.js
@@ -4,6 +4,7 @@
 var createError = require('http-errors')
 var router = require('express').Router()
 const User = require('../../../models/users')
+const Board = require('../../../models/boards')
 const crypto = require('crypto')
 
 //토큰을 위조한 것이 아니라 올바른 절차를 밞은 유저인지 확인하는 api
@@ -27,6 +28,9 @@ router.all('*', function (req, res, next) {
     next();
 });
 
+
+
+
 //이미지 수정
 router.put('/', (req, res, next) => {
     const _id = req.user._id;
@@ -43,11 +47,13 @@ router.put('/', (req, res, next) => {
 
 //유저 닉네임 비밀번호 수정,  회원탈퇴기능
 
-//매니저와 유저가 변경하는 것은 다르다는 것을 보여주기 위해 파일을 나눔(/manage/users, /user)
+
+
 router.put('/:_id', (req, res, next) => {
 
-    const _id = req.user._id
-
+    const _id = req.params._id
+    if (_id !== req.user._id)
+        if (req.user.lv) throw createError(403, '권한이 없습니다')
     //비밀번호가 있다면 비밀번호를 암호화 하고 db수정
     if (req.body.pwd) {
         const { pwd } = req.body
@@ -65,12 +71,22 @@ router.put('/:_id', (req, res, next) => {
 })
 
 router.delete('/:_id', (req, res, next) => {
-    const _id = req.user._id
-    User.deleteOne({ _id })
+    const _id = req.params._id
+    if (_id !== req.user._id)
+        if (req.user.lv) throw createError(403, '권한이 없습니다')
+    User.findById(_id)
         .then(r => {
-            res.send({ success: true, msg: r })
+
+            //유저보드를 먼저 삭제
+            return Board.deleteOne({ name: r.myBoard })
         })
-        .catch(e => {
+        .then(() => {
+
+            return User.deleteOne({ _id })
+        }
+        ).then(r => {
+            res.send({ success: true, msg: r, token: req.token })
+        }).catch(e => {
             res.send({ success: false, msg: e.message })
         })
 })
